feat: YAML-based named set definitions with engine integration (#60)
* feat: implement YAML-based set configuration with engine integration

Agent-Logs-Url: https://github.com/mazdakn/fwsim/sessions/7adb6e43-37b2-46a6-af5a-9927f9a9ca77


* docs: clarify OR semantics in set Match methods

Agent-Logs-Url: https://github.com/mazdakn/fwsim/sessions/7adb6e43-37b2-46a6-af5a-9927f9a9ca77


* refactor: change Set interface to Add(any)/Match(any) and revert typed helpers

Agent-Logs-Url: https://github.com/mazdakn/fwsim/sessions/68033a6c-9a1e-47af-b4f3-b6c6559ca14d


---------

+++ pkg/set/set_test.go
@@ -50,9 +50,9 @@ func TestPortSetAdd(t *testing.T) {
 
 	ps := NewPortSet()
 
-	ps.Add(80)
-	Expect(ps.Match(80)).To(BeTrue())
-	Expect(ps.Match(443)).To(BeFalse())
+	ps.Add(uint16(80))
+	Expect(ps.Match(uint16(80))).To(BeTrue())
+	Expect(ps.Match(uint16(443))).To(BeFalse())
 }
 
 func TestPortSetDelete(t *testing.T) {
@@ -60,13 +60,13 @@ func TestPortSetDelete(t *testing.T) {
 
 	ps := NewPortSet()
 
-	ps.Add(80)
-	ps.Add(443)
-	Expect(ps.Match(80)).To(BeTrue())
+	ps.Add(uint16(80))
+	ps.Add(uint16(443))
+	Expect(ps.Match(uint16(80))).To(BeTrue())
 
-	ps.Delete(80)
-	Expect(ps.Match(80)).To(BeFalse())
-	Expect(ps.Match(443)).To(BeTrue())
+	ps.Delete(uint16(80))
+	Expect(ps.Match(uint16(80))).To(BeFalse())
+	Expect(ps.Match(uint16(443))).To(BeTrue())
 }
 
 func TestPortSetMatch(t *testing.T) {
@@ -74,18 +74,18 @@ func TestPortSetMatch(t *testing.T) {
 
 	ps := NewPortSet()
 
-	Expect(ps.Match(80)).To(BeFalse())
+	Expect(ps.Match(uint16(80))).To(BeFalse())
 
-	ps.Add(80)
-	Expect(ps.Match(80)).To(BeTrue())
-	Expect(ps.Match(8080)).To(BeFalse())
+	ps.Add(uint16(80))
+	Expect(ps.Match(uint16(80))).To(BeTrue())
+	Expect(ps.Match(uint16(8080))).To(BeFalse())
 }
 
 func TestPortSetStringOnePort(t *testing.T) {
 	RegisterTestingT(t)
 
 	ps := NewPortSet()
-	ps.Add(80)
+	ps.Add(uint16(80))
 	Expect(ps.String()).To(Equal("80"))
 }
 
@@ -93,8 +93,8 @@ func TestPortSetStringMultiplePorts(t *testing.T) {
 	RegisterTestingT(t)
 
 	ps := NewPortSet()
-	ps.Add(443)
-	ps.Add(80)
+	ps.Add(uint16(443))
+	ps.Add(uint16(80))
 	Expect(ps.String()).To(Equal("{80,443}"))
 }
